fix(toolkit): centralize chat navigation
