+++ src/interfaces/assistants_web/src/components/Agents/AgentsSidePanel.tsx
@@ -1,8 +1,6 @@
 'use client';
 
 import { Transition } from '@headlessui/react';
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 import React, { ReactElement } from 'react';
 
 import { IconButton } from '@/components/IconButton';
@@ -10,13 +8,8 @@ import { Button, Icon, Logo, Text } from '@/components/Shared';
 import { Shortcut } from '@/components/Shortcut';
 import { env } from '@/env.mjs';
 import { useIsDesktop } from '@/hooks/breakpoint';
-import {
-  useAgentsStore,
-  useCitationsStore,
-  useConversationStore,
-  useParamsStore,
-  useSettingsStore,
-} from '@/stores';
+import { useNavigateToNewChat } from '@/hooks/chatRoutes';
+import { useAgentsStore, useSettingsStore } from '@/stores';
 import { cn } from '@/utils';
 
 /**
@@ -28,26 +21,13 @@ export const AgentsSidePanel: React.FC<React.PropsWithChildren<{ className?: str
   className = '',
   children,
 }) => {
-  const router = useRouter();
   const {
     agents: { isAgentsSidePanelOpen },
   } = useAgentsStore();
   const isDesktop = useIsDesktop();
   const isMobile = !isDesktop;
 
-  const { setEditAgentPanelOpen } = useAgentsStore();
-  const { resetConversation } = useConversationStore();
-  const { resetCitations } = useCitationsStore();
-  const { resetFileParams } = useParamsStore();
-
-  const handleNewChat = () => {
-    const url = '/';
-    setEditAgentPanelOpen(false);
-    resetConversation();
-    resetCitations();
-    resetFileParams();
-    router.push(url);
-  };
+  const navigateToNewChat = useNavigateToNewChat();
 
   return (
     <Transition
@@ -88,12 +68,12 @@ export const AgentsSidePanel: React.FC<React.PropsWithChildren<{ className?: str
             'justify-between gap-x-3': isMobile && isAgentsSidePanelOpen,
           })}
         >
-          <Link href="/">
+          <button onClick={() => navigateToNewChat()}>
             <Logo
               hasCustomLogo={env.NEXT_PUBLIC_HAS_CUSTOM_LOGO}
               includeBrandName={isAgentsSidePanelOpen}
             />
-          </Link>
+          </button>
 
           <ToggleAgentsSidePanelButton className="flex md:hidden" />
         </div>
@@ -110,7 +90,7 @@ export const AgentsSidePanel: React.FC<React.PropsWithChildren<{ className?: str
                 <Shortcut sequence={['⌘', '↑', 'N']} className="hidden group-hover:flex" />
               </div>
             }
-            onClick={handleNewChat}
+            onClick={() => navigateToNewChat()}
             tooltip="New chat"
             icon={<Icon name="add" kind="outline" className="dark:text-evolved-green-700" />}
           />
