+++ src/interfaces/assistants_web/src/components/Conversation/Header.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Transition } from '@headlessui/react';
-import { usePathname, useRouter } from 'next/navigation';
+import { usePathname } from 'next/navigation';
 import { useContext } from 'react';
 
 import { IconButton } from '@/components/IconButton';
@@ -12,24 +12,18 @@ import { WelcomeGuideTooltip } from '@/components/WelcomeGuideTooltip';
 import { ModalContext } from '@/context/ModalContext';
 import { useAgent } from '@/hooks/agents';
 import { useIsDesktop } from '@/hooks/breakpoint';
+import { useNavigateToNewChat } from '@/hooks/chatRoutes';
 import { WelcomeGuideStep, useWelcomeGuideState } from '@/hooks/ftux';
 import { useSession } from '@/hooks/session';
-import {
-  useAgentsStore,
-  useCitationsStore,
-  useConversationStore,
-  useParamsStore,
-  useSettingsStore,
-} from '@/stores';
+import { useAgentsStore, useConversationStore, useSettingsStore } from '@/stores';
 import { cn } from '@/utils';
 
 const useHeaderMenu = ({ agentId }: { agentId?: string }) => {
   const { open } = useContext(ModalContext);
   const {
     conversation: { id: conversationId },
-    resetConversation,
   } = useConversationStore();
-  const { resetCitations } = useCitationsStore();
+
   const { userId } = useSession();
   const { data: agent } = useAgent({ agentId });
   const isAgentCreator = userId === agent?.user_id;
@@ -42,18 +36,14 @@ const useHeaderMenu = ({ agentId }: { agentId?: string }) => {
     agents: { isEditAgentPanelOpen },
     setEditAgentPanelOpen,
   } = useAgentsStore();
-  const { resetFileParams } = useParamsStore();
-  const router = useRouter();
+
   const pathname = usePathname();
   const { welcomeGuideState, progressWelcomeGuideStep, finishWelcomeGuide } =
     useWelcomeGuideState();
 
+  const navigateToNewChat = useNavigateToNewChat();
   const handleNewChat = () => {
-    const url = agentId ? `/a/${agentId}` : pathname.includes('/a') ? '/a' : '/';
-    router.push(url, undefined);
-    resetConversation();
-    resetCitations();
-    resetFileParams();
+    navigateToNewChat(agentId);
   };
 
   const handleOpenShareModal = () => {
